+++ src/functions.h
@@ -7,11 +7,11 @@ struct Mapping
     char *addr = 0;
     int is_pmem = 0;
     int map_anon = 0; // MAP_ANONYMOUS (not backed by file)
-    long fsize = 0;
-    long buflen = 0;
+    uint64_t fsize = 0;
+    uint64_t buflen = 0;
     const char *fpath = "";
 };
 
-void get_bandwidth(int, int, int, Results &);
+void get_bandwidth(uint64_t, int, uint64_t, Results &);
 
 #endif
